create base interface for mining class
